Move database migration and fix database injection
